[Issue-4] Busca os detalhes do Hotel pelo seu ID
Controller para buscar os dados do hotel pelo seu ID

+++ src/main/java/br/com/hotel/http/dto/PriceDTO.java
@@ -9,8 +9,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 @Builder
@@ -22,8 +20,8 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 public class PriceDTO {
 
     @ApiModelProperty(notes = "Adult Price")
-    private BigDecimal adult;
+    private Double adult;
 
     @ApiModelProperty(notes = "Child Price")
-    private BigDecimal child;
+    private Double child;
 }
